refactor code to make router query the single source of truth
